refactoring

+++ src/main/java/stacs/wordle/scoreEngine/ScoreEngine.java
@@ -13,7 +13,7 @@ public class ScoreEngine {
         private static final float MAX_ATTEMPTS = 6;
         private static final float MAX_SCORE = 100;
         private static final float MIN_SCORE = 0;
-        private static final float SCORE_PER_ATTEMPT = 16.67F;
+        private static final double SCORE_PER_ATTEMPT = 16.67;
 
         private static final MainService MAIN_SERVICE = new MainService();
 
@@ -24,8 +24,8 @@ public class ScoreEngine {
         * @param attempts The number of attempts the player has made.
         * @return The score of the player.
         */
-        public static float calculateScore(int attempts) {
-            float score = MAX_SCORE - (attempts * SCORE_PER_ATTEMPT);
+        public static double calculateScore(int attempts) {
+            double score = MAX_SCORE - (attempts * SCORE_PER_ATTEMPT);
             if (score < MIN_SCORE) {
                 score = MIN_SCORE;
             }
